# doc/api.yaml
openapi: 3.0.2
info:
  title: WASAPhoto
  description: |
    OpenAPI document for the WASAPhoto application
  version: 1.1.0  
paths:   
...
  /users/{idUser}/followings:
    parameters:
      - $ref: "#/components/parameters/idUser"
    get:
      summary: List following users
      description: |
        List all the following of a specific user
      tags: ["follows"]
      operationId: listFollowings
      responses:
        "200": 
          description: Following returned
          content:
            application/json:
              schema:
                description: The list of following
                type: object
                properties:
                  followings:
                    description: The list of users followed by the specified user
                    type: array
                    minItems: 0
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/User"
        "400": 
          $ref: '#/components/responses/BadRequest'
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:  
...
    User:
      title: User
      description: |
        This object reperesents all the info related to a user
      type: object
      properties:
        id:
          description: The ID of the user
          type: integer
          readOnly: true
          example: 100
        name:
          description: User's name
          type: string
          pattern: '^.*?$'
          minLength: 2
          maxLength: 52
          example: Andrea
        surname:
          description: User's surname
          type: string
          pattern: '^.*?$'
          minLength: 2
          maxLength: 52
          example: Rossi
        username:
          $ref: "#/components/schemas/Username"
        postNo:
          description: The number of post
          type: integer
          minimum: 0
          example: 200
        followerNo:
          description: The number of follower
          type: integer
          minimum: 0
          example: 25
        followingNo:
          description: The number of followed
          type: integer
          minimum: 0
          example: 10
        followed:
          description: If the user is followed
          type: boolean
          example: false
    Username:
      title: Username
      description: This object represent the user name
      type: object
      properties:
        name:
          description: The username
          type: string
          pattern: '^[a-zA-Z0-9_\.#$]+$'
          minLength: 3
          maxLength: 16
          example: andremuso
...
    Message:
      title: Message
      description: This object represent a generic description of a message
      type: object
      readOnly: true
      properties:
        error:
          description: The message of the error
          type: string
          pattern: '\S.+'
          example: Username is already taken
          minLength: 0
          maxLength: 200
  parameters:
...
    idUser:
      name: idUser
      in: path
      schema: 
        $ref: "#/components/schemas/User/properties/id" 
      required: true
      description: The id of a user
...
  responses:
...
    BadRequest: # 400
      description: |
        The request sent to the server is invalid (e.g. malformed request syntax, invalid request message framing, ...)
      content:
        application/json:
          schema:
            type: object
            description: The message of the error
            properties:
              err:
                $ref: "#/components/schemas/Message"
    Unauthorized: # 401
      description: 
        The request was not successful because it lacks valid authentication credentials for the requested resource
      content:
        application/json:
          schema:
            type: object
            description: The message of the error
            properties:
              err:
                $ref: "#/components/schemas/Message"
...
    NotFound: # 404
      description: |
        The resource you are trying to access doesn't exist
      content:
        application/json:
          schema: 
            type: object
            description: The message of the error
            properties:
              err:
                $ref: "#/components/schemas/Message"
...
    InternalServerError:  #500
      description: The server encounter an unexpected error
      content:
        application/json:
          schema:
            type: object
            description: The message of the error
            properties:
              err:
                $ref: "#/components/schemas/Message"    
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

# Apply the security global, to all the operations
security:
  - bearerAuth: []
...